Get posts using listen for state change
